requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Graph Neural Networks
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration & Logging
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0

# Explainability (XAI)
captum>=0.6.0

# Utilities
joblib>=1.3.0
networkx>=3.1
h5py>=3.9.0

# Notebooks
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
